all generic now, better samples and readme

+++ Source/de/peote/events/PeoteEventDLL.hx
@@ -4,6 +4,7 @@ package de.peote.events;
  * ...
  * @author Sylvio Sell
  */
+
 @:generic
 class PeoteEventDLL<PARAM> extends PeoteDLL<PeoteEventNode<PARAM>>
 {
@@ -37,10 +38,10 @@ class PeoteEventDLL<PARAM> extends PeoteDLL<PeoteEventNode<PARAM>>
 
 	}
 	
-	public inline function listen(obs:PeoteDLL<PeoteDLLNode<PeoteEventNode<PARAM>>>, event_nr:Int, callback:Int->PARAM->Void, checkEventExists=true):Void
+	public inline function listen(obs:PeoteDLL<PeoteDLLNode<PeoteEventNode<PARAM>>>, event_nr:Int, callback:Int->PARAM->Void, checkEventExists:Bool):Void
 	{
 		// check if event is still listening !!!
-		if (checkEventExists) unlisten(obs, event_nr);  // TODO -> OPTIMIZE!!!!
+		if (checkEventExists) unlisten(obs, event_nr);
 		var obsNode:PeoteDLLNode<PeoteDLLNode<PeoteEventNode<PARAM>>> = obs.append(append( new PeoteEventNode(null, callback, event_nr) )); 
 		obsNode.node.node.listen = obsNode;		
 	}
@@ -50,14 +51,10 @@ class PeoteEventDLL<PARAM> extends PeoteDLL<PeoteEventNode<PARAM>>
 		node = head;
 		while (node != null)
 		{	
-			if (node.node.listen.dll == obs)
-			{
-				if (node.node.event_nr == event_nr)
-				{	
+			if (node.node.listen.dll == obs && node.node.event_nr == event_nr)
+			{	
 					obs.unlink(node.node.listen);
 					node = unlink(node);
-				}
-				else node = node.next;
 			}
 			else node = node.next;
 		}
